+++ commons/errors.go
@@ -52,6 +52,12 @@ func ValidateBusinessError(err error, entityType string, args ...any) error {
 			Title:      "Account Status Transaction Restriction",
 			Message:    "The current statuses of the source and/or destination accounts do not permit transactions. Change the account status(es) and try again.",
 		},
+		constant.ErrOverFlowInt64: Response{
+			EntityType: entityType,
+			Code:       constant.ErrOverFlowInt64.Error(),
+			Title:      "Overflow Error",
+			Message:    "The request could not be completed due to an overflow. Please check the values, and try again.",
+		},
 	}
 
 	if mappedError, found := errorMap[err]; found {